Fix Vite HMR breakage in Docker container
If server.hmr.host is unset, it defaults to localhost.  When Vite server
boots up, the link to Javascript assets compiled by Vite prepends with
localhost.  This breaks HMR when Vite dev server is running concurrently
with PHP development server in a Docker container hosted on a remote
Linux server.  The error might look like this:

"Cross-Origin Request Blocked: The Same Origin Policy disallows reading
the remote resource at http://localhost:<port number>/@vite/client.
(Reason: CORS request did not succeed). Status code: (null)."

This is a bug and it is 100% reproducible in the said environment.  Yet
it is very easy to fix by setting server.hmr.host to a valid link other
than localhost.  With that the browser load the hot module correctly.