Use named arguments and compact helper
Use named arguments for view helper.  This clearly shows what is the
name of the view that is used with the index action and what data is
passed from the controller to the view.

The order of these two named arguments doesn't matter any more, even
though we didn't use this feature.