Fix the non-constant expression as default value
We can't assign the result of a function call to the promoted parameter,
but we can use the scalar and set it to 0 for default value.  After its
initial assignment, we can override it with the result of the function
call in the constructor body.

It's an error to use a runtime expression where a compile-time constant
is required.

Update class definition document accordingly.